requirements.txt
asgiref==3.11.0
attrs==25.4.0
autobahn==25.12.2
Automat==25.4.16
cbor2==5.7.1
certifi==2025.11.12
cffi==2.0.0
channels==4.3.2
channels_redis==4.3.0
charset-normalizer==3.4.4
constantly==23.10.4
cryptography==46.0.3
daphne==4.2.1
Django==6.0
django-cors-headers==4.3.0
django-redis==6.0.0
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
h2==4.3.0
hpack==4.1.0
hyperframe==6.1.0
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
inflection==0.5.1
iniconfig==2.3.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
msgpack==1.1.2
packaging==25.0
pluggy==1.6.0
priority==1.3.0
psycopg2-binary==2.9.9
py-ubjson==0.16.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
Pygments==2.19.2
PyJWT==2.10.1
pyOpenSSL==25.3.0
pytest==9.0.2
pytest-asyncio==1.3.0
pytest-django==4.11.1
python-decouple==3.8
python-dotenv==1.0.0
pytz==2025.2
PyYAML==6.0.3
redis==5.0.0
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
service-identity==24.2.0
setuptools==80.9.0
sqlparse==0.5.5
Twisted==25.5.0
txaio==25.12.2
typing_extensions==4.15.0
ujson==5.11.0
uritemplate==4.2.0
urllib3==2.6.2
websockets==15.0.1
zope.interface==8.1.1
